+++ src/xrGame/console_commands_mp.cpp
@@ -1029,10 +1029,10 @@ public:
 		if (!tmp_sv_game) return;
 		string512 tmp_dest;
 		string512 filter_dest = "";
-		exclude_raid_from_args(args, tmp_dest, sizeof(tmp_dest));
+		exclude_raid_from_args(args, tmp_dest, (u32)sizeof(tmp_dest));
 		if (xr_strlen(tmp_dest))
 		{
-			sscanf_s(tmp_dest, "%s", filter_dest, sizeof(filter_dest));
+			sscanf_s(tmp_dest, "%s", filter_dest, (u32)sizeof(filter_dest));
 		}
 		tmp_sv_game->PrintBanList(filter_dest);
 		Level().Server->Print_Banned_Addreses();
@@ -1061,9 +1061,9 @@ public:
 		GameType[0]		=0;
 		
 		sscanf_s		(args,"%255s %255s %255s",
-			LevelName, sizeof(LevelName),
-			LevelVersion, sizeof(LevelVersion),
-			GameType, sizeof(GameType)
+			LevelName, (u32)sizeof(LevelName),
+			LevelVersion, (u32)sizeof(LevelVersion),
+			GameType, (u32)sizeof(GameType)
 		);
 
 		EGameIDs GameTypeID = ParseStringToGameType(GameType);
@@ -1157,8 +1157,8 @@ public:
 		LevelName[0]	=	0;
 		LevelVersion[0] =	0;
 		sscanf_s		(args,"%255s %255s",
-			LevelName, sizeof(LevelName),
-			LevelVersion, sizeof(LevelVersion)
+			LevelName, (u32)sizeof(LevelName),
+			LevelVersion, (u32)sizeof(LevelVersion)
 		);
 
 		string1024		argsNew;
